+++ inst/htmlwidgets/phylocanvas.js
@@ -15,12 +15,15 @@ HTMLWidgets.widget({
       renderValue: function(x) {
         //pass params globally
 
+        // clear the sub nodes. Necessary for shiny apps
+        while (el.firstChild)
+          el.removeChild(el.firstChild);
+
         var tree = Phylocanvas.createTree(el, x.config);
 
         //set global variables
         tree.setTreeType(x.treetype);
         tree.setNodeSize(x.nodesize);
-        tree.setTextSize(x.textsize);
         tree.lineWidth = x.linewidth;
         tree.showLabels = x.showlabels;
         tree.alignLabels = x.alignlabels;
@@ -30,11 +33,13 @@ HTMLWidgets.widget({
 
         // loading will draw the tree
         tree.load(x.tree);
+        
+        tree.setTextSize(x.textsize);
 
         // apply node styles after the tree is loaded
-        if (x.nodestyles) {
+        if (x.nodestyles !== null) {
           Object.keys(x.nodestyles).forEach(function(key){
-          var leaf = tree.findLeaves(key)[0];
+          var leaf = tree.findLeaves("^" + key + "$")[0];
           leaf.setDisplay( x.nodestyles[key] );
           leaf.highlighted = x.nodestyles[key].highlighted;
           });