+++ NEWS.md
@@ -1,3 +1,12 @@
+# 0.1.2
+
+* fix text size issue
+
+# 0.1.1
+
+* fix child node issue for correct behavior in Shiny
+* remove unused phylocanvas dependencies
+
 # phylocanvas 0.1.0
 
 * first release to CRAN.